Ensure atomicity of MySQL compare-and-swap
MySQL transactions do not by default provide serializable snapshot
isolation. As such, simply wrapping the component steps of a
compare-and-swap operation in a transaction is insufficient to satisfy
the invariants required of such an operation. Acquiring an update lock
as part of the initial SELECT query resolves this by forcing concurrent
compare-and-swap operations to operate serially.

Fixes #17