feat: adds a MongoDB connector
This commit adds a MongoDB connector.

A couple of things to note:
- To enable transactions the database must a replica set,
I made the connection fail if the condition was not met.

MongoDB is a document database, thus concepts translate as:
bucket/table = collection
key = document
value = document value

Note: mongo will automatically create a collection
if a operation is attempted on one that does not exist.
For example: if you delete a table, then call list to
check for the contents of the table it will be recreated.

closes smallstep/certificates#141